+++ src/components/certification/index.ts
@@ -1,0 +1,5 @@
+export { FacilityCertifierSection } from "./facility-certifier-section";
+export {
+  FacilityCertifierDialog,
+  UnlinkConfirmDialog,
+} from "./facility-certifier-dialog";